Example of Reading SMD530CT by Modbus only
This program is a example for read a SDM530CT by modbus. The fucntion use is ReadInputRegisters take account in modbus protocol this is a function code 0x04. Registers reading are RMS voltage phase 1 address 0x00 two bytes, RMS current in phase 1 address 0x06 two bytes, RMS active power phase 1 address 0x0C, the register is the power factor of grid  with address 0x1E, the phase angle 1 is 0x24 register, and finally the last register reading in this example is the frequency in the register 0x24.